# specs/001-product-catalog/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Product Catalog API
  version: 1.2.0
servers:
  - url: http://localhost:3000
paths:
  /api/auth/login:
    post:
      summary: Admin login
      description: Exchanges valid admin credentials for a short-lived JWT.
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
            examples:
              loginExample:
                value:
                  username: admin
                  password: secret
      responses:
        '200':
          description: Successful authentication response containing JWT and expiry.
          content:
            application/json:
              schema:
                type: object
                required: [token, expiresAt]
                properties:
                  token:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                invalidCredentials:
                  value:
                    error: invalid_credentials
                    message: Invalid admin credentials
  /api/products:
    get:
      summary: List products
      description: Returns up to 100 products (hard cap). Optional text search and category filtering.
      operationId: listProducts
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Case-insensitive substring matched against name and description.
        - in: query
          name: categoryId
          schema:
            type: string
          description: Filter products by category id.
      responses:
        '200':
          description: OK — array of Product (max 100 items)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Server error
    post:
      summary: Create product
      description: Creates a product (admin only).
      security:
        - bearerAuth: []
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, description, price]
              properties:
                name: { type: string }
                description: { type: string }
                price: { type: number }
                categoryId: { type: string }
                stock: { type: integer, minimum: 0 }
            examples:
              createProduct:
                value:
                  name: Sample Product A
                  description: Example A description
                  price: 10
                  categoryId: 22222222-2222-4222-8222-222222222222
                  stock: 5
      responses:
        '201':
          description: Created product.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400': { description: Validation error }
        '401':
          description: Missing or invalid admin authentication.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StandardError' }
        '403':
          description: Authenticated but lacks required admin role.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StandardError' }
  /api/products/{id}:
    get:
      summary: Get product by id
      operationId: getProduct
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Product found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404': { description: Not found }
    put:
      summary: Update product
      description: Updates product fields (admin only).
      security: [ { bearerAuth: [] } ]
      operationId: updateProduct
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                price: { type: number }
                categoryId: { type: string }
                stock: { type: integer, minimum: 0 }
      responses:
        '200': { description: Updated product, content: { application/json: { schema: { $ref: '#/components/schemas/Product' } } } }
        '400': { description: Validation error }
        '401': { description: Admin auth required, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
        '403': { description: Forbidden admin role, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
        '404': { description: Not found }
    delete:
      summary: Delete product
      description: Deletes product (admin only).
      security: [ { bearerAuth: [] } ]
      operationId: deleteProduct
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
        '401': { description: Admin auth required, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
        '403': { description: Forbidden admin role, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
        '404': { description: Not found }
  /api/categories:
    get:
      summary: List categories
      operationId: listCategories
      responses:
        '200':
          description: OK — array of Category
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Category' }
    post:
      summary: Create category
      description: Creates a category (admin only).
      operationId: createCategory
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryInput' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Category' } } } }
        '400': { description: Validation error }
        '401': { description: Admin auth required, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
        '403': { description: Forbidden admin role, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
  /api/categories/{id}:
    get:
      summary: Get category by id
      operationId: getCategory
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Category' } } } }
        '404': { description: Not found }
    put:
      summary: Update category
      description: Updates category name (admin only).
      operationId: updateCategory
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryInput' }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Category' } } } }
        '400': { description: Validation error }
        '401': { description: Admin auth required, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
        '403': { description: Forbidden admin role, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
        '404': { description: Not found }
    delete:
      summary: Delete category (blocked if products exist)
      description: Deletes category if no products reference it (admin only). Returns conflict if referenced.
      operationId: deleteCategory
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
        '409': { description: Conflict — category has assigned products }
        '401': { description: Admin auth required, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
        '403': { description: Forbidden admin role, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
        '404': { description: Not found }
  /api/orders:
    post:
      summary: Submit order
      description: Creates order snapshot then atomically decrements stock. Rejects entirely if any item lacks stock.
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderInput' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '400': { description: Validation error }
        '409': { description: Conflict — insufficient stock }
  /api/orders/{id}:
    get:
      summary: Get order by id
      operationId: getOrder
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '404': { description: Not found }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    StandardError:
      type: object
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }
    Product:
      type: object
      required: [id, name, description, price, stock, imageUrl, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        price: { type: number, format: double }
        categoryId: { type: string, nullable: true }
        stock: { type: integer, minimum: 0 }
        imageUrl: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Category:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CategoryInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
    OrderItem:
      type: object
      required: [productId, name, price, quantity]
      properties:
        productId: { type: string }
        name: { type: string }
        price: { type: number }
        quantity: { type: integer, minimum: 1 }
    Order:
      type: object
      required: [id, items, total, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
        total: { type: number }
        status: { type: string, enum: [submitted] }
        createdAt: { type: string, format: date-time }
    OrderInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId: { type: string }
              quantity: { type: integer, minimum: 1 }